requirements.txt
Django~=4.0.8
djangorestframework==3.14.0

pika==1.3.1

#testing - factory for dummy
factory-boy==3.2.1
